Fixed an issue with loading the caching strategy from configuraiton
It wasn't updating the QueryManager so it was using the
NoCachingStrategy even when the CachingStrategy got updated, the
QueryManager didn't.

Changed exceptions in configuraiton loading to
ConfigurationErrorsException instead of ArgumentException.

Added web.config transformation files but haven't got them working yet
when creating a new nuget package, that comes next.

+++ nuget/nuget-build.ps1
@@ -19,33 +19,39 @@ $new_notes = Read-Host "Any release notes?"
 
 Write-Host ""
 
-$directories = Get-Directory $(Get-Location), "SharpRepository.*" -recurse
-
+$directories = Get-Directory $(Get-Location), "SharpRepository.*"
+$uniqueDirectories = @()
 foreach($directory in $directories)
 {
 	$dir_fullname = $directory.fullname
-
-	Write-Host "Processing " $dir_fullname
 	
-	$template_path = ($directory.fullname + "\" + $directory + ".nuspec.template")
-	$nuspec_path = ($directory.fullname + "\" + $directory + ".nuspec")
-	
-	# update the version number
-	((Get-Content $template_path) -creplace $version_holder, ("<version>" + $new_version + "</version>")) | Set-Content $nuspec_path
-    ((Get-Content $nuspec_path) -creplace $notes_holder, ("<releaseNotes>" + $new_notes + "</releaseNotes>")) | Set-Content $nuspec_path
-	
-	# copy the nuspec file over to the main project directory because I only know how to create the nuget package from that diretory, but that's just me i think :)
-	Copy-Item $nuspec_path "..\$directory\$directory.nuspec"
-	
-	# build the nuget package
-	#& "C\NuGet\nuget.exe pack ..\$directory\$directory.csproj -Prop Configuration=Release"
-	Invoke-Expression "& `"nuget.exe`" pack ..\$directory\$directory.csproj -Prop Configuration=Release"
-	
-	# move the nuget package back to the nuget directories
-	Move-Item "$cur_directory\$directory.$new_version.nupkg" "$dir_fullname\$directory.$new_version.nupkg"
-	
-	# delete the nuspec file from the project directory
-	Remove-Item "..\$directory\$directory.nuspec"
+	if (!($uniqueDirectories -contains $dir_fullname) )
+	{ 
+		Write-Host "Processing " $dir_fullname
+		
+		$template_path = ($directory.fullname + "\" + $directory + ".nuspec.template")
+		$nuspec_path = ($directory.fullname + "\" + $directory + ".nuspec")
+		
+		# update the version number
+		((Get-Content $template_path) -creplace $version_holder, ("<version>" + $new_version + "</version>")) | Set-Content $nuspec_path
+		((Get-Content $nuspec_path) -creplace $notes_holder, ("<releaseNotes>" + $new_notes + "</releaseNotes>")) | Set-Content $nuspec_path
+		
+		# copy the nuspec file over to the main project directory because I only know how to create the nuget package from that diretory, but that's just me i think :)
+		Copy-Item $nuspec_path "..\$directory\$directory.nuspec"
+		
+		# build the nuget package
+		#& "C\NuGet\nuget.exe pack ..\$directory\$directory.csproj -Prop Configuration=Release"
+		Invoke-Expression "& `"nuget.exe`" pack ..\$directory\$directory.csproj -Prop Configuration=Release"
+		
+		# move the nuget package back to the nuget directories
+		Move-Item "$cur_directory\$directory.$new_version.nupkg" "$dir_fullname\$directory.$new_version.nupkg"
+		
+		# delete the nuspec file from the project directory
+		Remove-Item "..\$directory\$directory.nuspec"
+		
+		# add the full path to the list of already processed directories, this is because the array for this loop has dups
+		$uniqueDirectories += $directory.fullname
+	}
 }
 
 Read-Host -prompt "Finished.  Press Enter key to exit"
